website/requirements.txt
Flask
pymongo
pyyaml
pika